starting project sanitization
